Move submission simulation to separate file
